agregando la funcionalidad para pasar a la siguiente pregunta con un boton
